reading of map and SFINAE tricks
Reading of the map is now a part of config file.
User now can choose any IO files on his system to be used by the program.
This code uses std::filesystem library so should work fine on any OS.
Code is tolerant to the changes of containers used to store map elements. SFINAE complile-time tricks are implemented to ensure correct container behavior.